close #2 add crud in cruds
Feat:
	- add package cruds
	- add modules in this package for splitting

Changes:
	- refact comments for pep8
