+++ routers/mission.py
@@ -2,18 +2,19 @@ from fastapi import Depends, APIRouter
 from sqlalchemy.orm import Session
 from depedency import get_db
 from appsql import schemas
-import crud
-
+from cruds import crud_mission
 
 router = APIRouter()
 
-#Fonction POST
+
+# POST Function
 @router.post("/missions/", response_model=schemas.Mission)
 def create_mission(mission: schemas.MissionCreate, db: Session = Depends(get_db)):
-    return crud.create_mission(db=db, mission=mission)
+    return crud_mission.create_mission(db=db, mission=mission)
 
-#Fonction GET
+
+# GET Function
 @router.get("/missions/", response_model=list[schemas.Mission])
 def read_mission(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    mission = crud.get_mission(db, skip=skip, limit=limit)
+    mission = crud_mission.get_mission(db, skip=skip, limit=limit)
     return mission
